requirements.txt
# --- For-Bot-Working --------- #
pyrofork
TgCrypto
pyromod==1.5



# --- For-Database ------------ #
pymongo
dnspython



# --- For-Web-Response ------- #
aiohttp


# --- For-env-variables ------- #
python-dotenv

requests
